Punto 7.2 90%

+++ cliente/src/main/java/org/una/examen/cliente/service/DistritoService.java
@@ -85,7 +85,7 @@ public class DistritoService {
         }
     }
     
-    public Respuesta delete(Integer id){
+    public Respuesta delete(Long id){
         try{
             Map<String, Object> parametros = new HashMap<>();
             parametros.put("id", id);
